requirements.txt
backports.zoneinfo;python_version<"3.9"
matplotlib==3.5.3 
numpy==1.21.5
opencv-contrib-python-headless==4.6.0.66
streamlit==1.19.0
pandas==1.5.3
pillow==9.4.0
plotly==5.13.1
scikit-learn==1.2.1 
